+++ src/main/java/dinkplugin/DinkPluginConfig.java
@@ -178,14 +178,14 @@ public interface DinkPluginConfig extends Config {
     )
     String chatSection = "Custom Chat Messages";
 
-    /*
+    
     @ConfigSection(
         name = "Leagues",
         description = "Settings for notifying when you complete league tasks, unlock areas, and redeem relics",
         position = 200,
         closedByDefault = true
     )
-     */
+     
     String leaguesSection = "Leagues";
 
     @ConfigSection(
@@ -405,7 +405,8 @@ public interface DinkPluginConfig extends Config {
     @ConfigItem(
         keyName = "ignoreSeasonalWorlds",
         name = "Ignore Seasonal Worlds",
-        description = "Whether to suppress notifications that occur on seasonal worlds like Leagues.",
+        description = "Whether to suppress notifications that occur on seasonal worlds like Leagues.<br/>" +
+            "Note: the Leagues-specific notifier uses an independent config toggle",
         position = 1015,
         section = advancedSection
     )
@@ -687,8 +688,7 @@ public interface DinkPluginConfig extends Config {
         description = "If non-empty, Leagues messages are sent to this URL, instead of the primary URL.<br/>" +
             "Note: this only applies to the Leagues notifier, not every notifier in a seasonal world",
         position = -1,
-        section = webhookSection,
-        hidden = true
+        section = webhookSection
     )
     default String leaguesWebhook() {
         return "";
@@ -2011,8 +2011,7 @@ public interface DinkPluginConfig extends Config {
         name = "Enable Leagues",
         description = "Enable notifications upon various leagues events",
         position = 200,
-        section = leaguesSection,
-        hidden = true
+        section = leaguesSection
     )
     default boolean notifyLeagues() {
         return false;
@@ -2023,8 +2022,7 @@ public interface DinkPluginConfig extends Config {
         name = "Send Image",
         description = "Send image with the notification",
         position = 201,
-        section = leaguesSection,
-        hidden = true
+        section = leaguesSection
     )
     default boolean leaguesSendImage() {
         return true;
@@ -2035,8 +2033,7 @@ public interface DinkPluginConfig extends Config {
         name = "Send Area Unlocks",
         description = "Send notifications upon area unlocks",
         position = 202,
-        section = leaguesSection,
-        hidden = true
+        section = leaguesSection
     )
     default boolean leaguesAreaUnlock() {
         return true;
@@ -2047,8 +2044,7 @@ public interface DinkPluginConfig extends Config {
         name = "Send Relic Unlocks",
         description = "Send notifications upon relic unlocks",
         position = 203,
-        section = leaguesSection,
-        hidden = true
+        section = leaguesSection
     )
     default boolean leaguesRelicUnlock() {
         return true;
@@ -2059,8 +2055,7 @@ public interface DinkPluginConfig extends Config {
         name = "Send Completed Tasks",
         description = "Send notifications upon completing a task",
         position = 204,
-        section = leaguesSection,
-        hidden = true
+        section = leaguesSection
     )
     default boolean leaguesTaskCompletion() {
         return true;
@@ -2071,8 +2066,7 @@ public interface DinkPluginConfig extends Config {
         name = "Task Min Difficulty",
         description = "The minimum tier of a task for a notification to be sent",
         position = 205,
-        section = leaguesSection,
-        hidden = true
+        section = leaguesSection
     )
     default LeagueTaskDifficulty leaguesTaskMinTier() {
         return LeagueTaskDifficulty.EASY;